allwinner: call PMIC setup code
As the ATF may need to do some power initialization on Allwinner
platform with AXP PMICs, call the PMICsetup code in BL31.

To prevernt wrong PMIC initialization sequence being passed to the
board, it's only called when DT name is determined.

Stub of PMIC setup code is added, to prevent undefined reference.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>